requirements.txt
Flask
gunicorn
APScheduler
Flask-APScheduler
Flask-Login
Flask-Mail
Flask-SQLAlchemy
google-api-core
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
cachetools
requests
Flask-OAuthlib
Authlib
pyrebase4
flask-googlemaps
python-dotenv
pyodbc